+++ packages/elements-web-components/.storybook/config.js
@@ -1,3 +1,0 @@
-import '@stoplight/storybook-config/config';
-
-import '@stoplight/elements/src/styles/elements.scss';