+++ packages/elements-web-components/.storybook/preview.js
@@ -1,0 +1,1 @@
+import '@stoplight/elements/src/styles/elements-scoped.scss';