+++ packages/elements/src/__stories__/index.ts
@@ -1,6 +1,8 @@
+// Public Components
+import './components/Api';
+import './components/StoplightProject';
+// Internal Components
 import './components/Dependencies';
 import './components/Docs';
 import './components/TableOfContents';
 import './components/TryIt';
-import './components/Api';
-import './components/StoplightProject';