+++ packages/elements/src/containers/API.tsx
@@ -1,28 +1,30 @@
-import { NodeType } from '@stoplight/types';
-import { FAIcon, NonIdealState, TableOfContents } from '@stoplight/ui-kit';
+import { generateToC } from '@stoplight/elements-utils';
+import { FAIcon, NonIdealState } from '@stoplight/ui-kit';
 import axios from 'axios';
 import * as React from 'react';
 import { useLocation } from 'react-router-dom';
 import useSwr from 'swr';
 
-import { Docs, DocsSkeleton } from '../components/Docs';
-import { Row } from '../components/TableOfContents/Row';
-import { TryIt } from '../components/TryIt';
-import { TryItHeader } from '../components/TryIt/header';
+import { SidebarLayout } from '../components/API/SidebarLayout';
+import { StackedLayout } from '../components/API/StackedLayout';
+import { DocsSkeleton } from '../components/Docs/Skeleton';
 import { withRouter } from '../hoc/withRouter';
-import { useBundledData } from '../hooks/useBundledData';
 import { useParsedValue } from '../hooks/useParsedValue';
-import { useTocContents } from '../hooks/useTocContents';
 import { withStyles } from '../styled';
-import { IAPI } from '../types';
-import { computeTocTree, isOas2, isOas3, isOperation, IUriMap, MODEL_REGEXP, OPERATION_REGEXP } from '../utils/oas';
+import { LinkComponentType, RoutingProps } from '../types';
+import { computeNodeData, isOas2, isOas3, IUriMap } from '../utils/oas';
 import { computeOas2UriMap } from '../utils/oas/oas2';
 import { computeOas3UriMap } from '../utils/oas/oas3';
-import { InlineRefResolverProvider } from './Provider';
 
 const fetcher = (url: string) => axios.get(url).then(res => res.data);
 
-const APIImpl = withRouter<IAPI>(({ apiDescriptionUrl, linkComponent: LinkComponent }) => {
+export interface APIProps extends RoutingProps {
+  apiDescriptionUrl: string;
+  linkComponent?: LinkComponentType;
+  layout?: 'sidebar' | 'stacked';
+}
+
+const APIImpl = withRouter<APIProps>(function API({ apiDescriptionUrl, linkComponent, layout }) {
   const { pathname } = useLocation();
 
   const { data, error } = useSwr(apiDescriptionUrl, fetcher);
@@ -34,7 +36,6 @@ const APIImpl = withRouter<IAPI>(({ apiDescriptionUrl, linkComponent: LinkCompon
   }, [error]);
 
   const document = useParsedValue(data);
-  const showTryIt = isOperation(pathname);
 
   const uriMap = React.useMemo(() => {
     let map: IUriMap = {};
@@ -50,20 +51,9 @@ const APIImpl = withRouter<IAPI>(({ apiDescriptionUrl, linkComponent: LinkCompon
     return map;
   }, [document]);
 
-  const tree = computeTocTree(uriMap);
-  const contents = useTocContents(tree).map(item => ({
-    ...item,
-    isActive: item.to === pathname,
-    isSelected: item.to === pathname,
-  }));
-
-  const nodeType = MODEL_REGEXP.test(pathname)
-    ? NodeType.Model
-    : OPERATION_REGEXP.test(pathname)
-    ? NodeType.HttpOperation
-    : NodeType.HttpService;
+  const nodes = computeNodeData(uriMap);
+  const tree = generateToC(nodes);
   const nodeData = uriMap[pathname] || uriMap['/'];
-  const bundledNodeData = useBundledData(nodeType, nodeData, { baseUrl: apiDescriptionUrl });
 
   if (error) {
     return (
@@ -83,26 +73,18 @@ const APIImpl = withRouter<IAPI>(({ apiDescriptionUrl, linkComponent: LinkCompon
 
   return (
     <div className="APIComponent flex flex-row">
-      <TableOfContents
-        contents={contents}
-        rowComponent={Row}
-        rowComponentExtraProps={{ pathname, linkComponent: LinkComponent }}
-      />
-      <div className="flex-grow p-5">
-        <div className="flex">
-          <InlineRefResolverProvider document={document}>
-            <Docs className="px-10" nodeData={bundledNodeData} nodeType={nodeType} />
-            {showTryIt && (
-              <div className="w-2/5 border-l relative">
-                <div className="absolute inset-0 overflow-auto px-10">
-                  <TryItHeader />
-                  <TryIt nodeType={nodeType} nodeData={bundledNodeData} />
-                </div>
-              </div>
-            )}
-          </InlineRefResolverProvider>
-        </div>
-      </div>
+      {layout === 'stacked' ? (
+        <StackedLayout uriMap={uriMap} tree={tree} apiDescriptionUrl={apiDescriptionUrl} document={document} />
+      ) : (
+        <SidebarLayout
+          pathname={pathname}
+          uriMap={uriMap}
+          tree={tree}
+          linkComponent={linkComponent}
+          apiDescriptionUrl={apiDescriptionUrl}
+          document={document}
+        />
+      )}
     </div>
   );
 });