+++ packages/elements/src/containers/Provider.tsx
@@ -5,7 +5,6 @@ import { get, isObject } from 'lodash';
 import * as React from 'react';
 import { Client, Provider as UrqlProvider } from 'urql';
 
-import { defaultComponents } from '../hooks/useComponents';
 import { useUrqlClient } from '../hooks/useUrqlClient';
 import { NodeIconMapping } from '../types';
 
@@ -22,10 +21,7 @@ export interface IActiveInfo {
   node?: string;
 }
 
-export const ComponentsContext = createNamedContext<IComponentMapping | undefined>(
-  'ComponentsContext',
-  defaultComponents,
-);
+export const ComponentsContext = createNamedContext<IComponentMapping | undefined>('ComponentsContext', undefined);
 
 export const InlineRefResolverContext = React.createContext<SchemaTreeRefDereferenceFn | undefined>(void 0);